requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymongo==4.6.0
dnspython==2.4.2
requests==2.31.0
python-decouple==3.8
mongoengine==0.27.0
django-cors-headers==4.3.1

